requirements.txt
autogen-agentchat
autogen-ext[azure]
aiohttp
tiktoken
openai
python-dotenv